While thinking about adding some icon or drawing to mark and fill the void at the top of the page where a Greek phrase is displayed, an initial consideration was to use some vector graphics icons depicting a scroll, one of an open scroll for the top of the page, and one of a rolled up scroll to use as an exit button (respectively Scroll by stefano corradetti and Scroll by JMA, both from the Noun Project).
However, trying to consider things from a media ecology point of view, it appeared that adding an icon, that is ideogram, would probably alter the sensorial balance and proportions of the app. Drawings would be a polluting noise distracting and detracting from the greek phrases.

Right now the app stimulates primarily two senses: the hearing, mainly, focusing on impressing a sound memory of minimal building blocks of the Greek language (micro phrases of about three words, usually); and, secondarily, the sight, but stripped down to bare bones level (black-and-white text, from a phonetic alphabet), in order to be only in a subaltern and ancillary role to the hearing, that is, serving to complete the understanding of the words just heard. This in consideration of McLuhan's observations in The Gutenberg Galaxy.

Then, if using only alphanumeric symbols seemed to be the only option, came up the possibility of putting a Greek letter at the top of each screen, serving both as a decoration of the blank space at the top and as a page number, since in Greek letters served also as numerals, therefore displaying a lower case alpha for the first page and so on. But that would also have conveyed an additional information (the number of the page we are on, which is already displayed in a widget in the bottom right corner of the page), and been misleading and distracting from the the Greek text.

Then came the idea of using some of the accents or diacritic symbols, that is either the aspiration mark, the acute or grave accent, and finally it all settled up on using the circumflex accent, or tilde, which also resembles a basic fleuron book ornament.